Allow the `data' argument to put() etc. to be a File
and if it is, use the given argument to upload from a file pointer.
This allows most methods to perform requests with bodies sourced from files,
bypassing the MRI heap. We leave the old code path in place to allow multipart
uploads to work, and we can't really remove the "filename" argument either, just
yet.

Also fixes a potential bug where the input file size was not set for the upload